requirements.txt
Django==1.10.4
django-secret-key==1.0.1
future==0.16.0
line-bot-sdk==1.0.2
requests==2.12.3
gunicorn==19.0.0
